backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9

numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==3.0.1

pypdf==4.2.0
markdown==3.6

openai==1.37.0   # optional if you want LLM answers
tiktoken==0.7.0  # optional token counting
